fix script issues in migration area and reorder down migration for 1.0.9
restore_db.sh failed to check it had been given a dump file to excute.

wipeDatabase.sh was going to always wipe db_sosm, can now be passed
in environment variable.
This script should probably verify as this is potentially
non-recoverable, in addition it assumes you are running as root to
access the DB.

1.0.9 down migration didn't undo all of it's actions of up, so added
  extra steps, but also reversed the existing order so that it was
  effectively undoing the changes in the reverse of what they'd been
  applied. In a complex migration this is probably a better format.